removed old crop and replaced with new cropv2
